clearclause-legal-assistant/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai
python-dotenv
langchain-google-genai
langchain
langchain-community
python-docx
faiss-cpu
PyMuPDF
tiktoken
pydantic
scikit-learn
